Move CalculateTaskGraphOperationResult
- from org.gradle.initialization to org.gradle.execution.taskgraph

+review REVIEW-6498
