+++ subprojects/core/src/main/java/org/gradle/execution/taskgraph/CalculateTaskGraphDescriptor.java
@@ -24,7 +24,7 @@ import java.util.Set;
 /*
  * Operation Descriptor for Calculating the task graph of a build.
  * This class is intentionally internal and consumed by the build scan plugin.
- * 
+ *
  * @since 4.0
  */
 public class CalculateTaskGraphDescriptor {
